+++ src/graphs/types/DirectedWeightedGraphAsMatrix.java
@@ -3,7 +3,7 @@ package types;
 import java.io.File;
 import java.io.IOException;
 
-public class DirectedWeightedGraphAsMatrix extends AbstractWeightedGraphAsMatrix implements DirectedGraph{
+public class DirectedWeightedGraphAsMatrix extends AbstractWeightedGraphAsMatrix implements DirectedGraph {
 
 	public DirectedWeightedGraphAsMatrix(Integer numberOfVertices) {
 		super(numberOfVertices);
@@ -14,7 +14,7 @@ public class DirectedWeightedGraphAsMatrix extends AbstractWeightedGraphAsMatrix
 	}
 
 	@Override
-	public void addEdgeBetween(Integer v, Integer w, Integer weight) {
+	public void addEdgeBetween(Integer v, Integer w, Double weight) {
 		edges[v][w] = weight.doubleValue();
 		numberOfEdges++;
 	}
