+++ src/graphs/types/WeightedGraph.java
@@ -1,11 +1,11 @@
 package types;
 
 public interface WeightedGraph extends Graph {
-	
-	void addEdgeBetween(Integer v, Integer w, Integer weight);
-	
-	Integer getWeightOfEdge(Integer v, Integer w);
-	
+
+	void addEdgeBetween(Integer v, Integer w, Double weight);
+
+	Double getWeightOfEdge(Integer v, Integer w);
+
 	boolean hasNegativeEdge();
-	
+
 }
